Release05

+++ Projects/App/Sources/Services/APIs/Models/StationLocation.swift
@@ -11,7 +11,7 @@ struct StationLocation: FireStoreCodable {
     /// 지하철 호선
     let route: String
     /// 지하철 역 코드명
-    let statnId: Int32
+    let statnId: Int64
     /// 지하철 역명
     let statnNm: String
 }
